src/pyproject.toml
[tool.poetry]
name = "web-raider"
version = "0.0.2"
description = "Web Raider"
authors = [
    "Prannaya <prannayagupta@gmail.com>",
    "En Hao <enhaotew@gmail.com>"
]
readme = "README.md"
package-mode = true
repository = "https://github.com/ThePyProgrammer/web-raider"

[tool.poetry.dependencies]
python = "^3.9"
newspaper3k = "^0.2.8"
googlesearch-python = "^1.2.5"
lxml = {extras = ["html-clean"], version = "^5.3.0"}
requests = "^2.32.3"
scikit-learn = "^1.5.1"
bs4 = "^0.0.2"
litellm = "^1.46.6"
boto3 = "^1.35.22"
pandas = "^2.2.2"
fastapi = "^0.115.0"
websockets = "^13.1"
uvicorn = "^0.30.6"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4"
pytest-sugar = "*"
pytest-cov = "*"
black = "*"
mypy = "*"
ruff = "*"
isort = "*"
tox = "^4.11.4"
ipykernel = "^6.29.5"

[tool.poetry.scripts]
web-raider-conn-mann = "web_raider.connection_manager:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
